home work 9 fixed
